% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_translation.R
\name{add_translation}
\alias{add_translation}
\title{Add a translation to a dictionary}
\usage{
add_translation(dictionary, word_nl, word_en, word_es)
}
\arguments{
\item{dictionary}{A data frame with the columns nl, en and es.}

\item{word_nl}{Dutch word.}

\item{word_en}{English word.}

\item{word_es}{Spanish word.}
}
\value{
A data frame with the new translation added.
}
\description{
This function adds a new translation row to an existing dictionary.
}
\examples{
dict <- data.frame(
  nl = "hond",
  en = "dog",
  es = "perro",
  stringsAsFactors = FALSE
)
add_translation(dict, "boek", "book", "libro")

}
