% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_languages.R
\name{available_languages}
\alias{available_languages}
\title{Show available languages}
\usage{
available_languages()
}
\value{
A character vector with available language codes.
}
\description{
This function returns the language codes that are available in the package.
}
\examples{
available_languages()

}
