% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_sentence.R
\name{translate_sentence}
\alias{translate_sentence}
\title{Translate a simple sentence}
\usage{
translate_sentence(sentence, from = "nl", to = "en")
}
\arguments{
\item{sentence}{A character string with the sentence to translate.}

\item{from}{The language code of the input sentence. Options are "nl", "en", or "es".}

\item{to}{The language code of the output sentence. Options are "nl", "en", or "es".}
}
\value{
A translated sentence as a character string.
}
\description{
This function translates a short sentence word by word using the built-in dictionary.
}
\examples{
translate_sentence("ik hou van koffie", from = "nl", to = "en")

}
