% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_word.R
\name{translate_word}
\alias{translate_word}
\title{Translate a single word}
\usage{
translate_word(word, from = "nl", to = "en")
}
\arguments{
\item{word}{A character string with the word to translate.}

\item{from}{The language code of the input word. Options are "nl", "en", or "es".}

\item{to}{The language code of the output word. Options are "nl", "en", or "es".}
}
\value{
A translated word as a character string. If the word is not found, the function returns NA.
}
\description{
This function translates one word between Dutch, English and Spanish.
}
\examples{
translate_word("hond", from = "nl", to = "en")
translate_word("coffee", from = "en", to = "nl")

}
